Move load_user_from_request to proper place
