+++ saleor/core/auth_backend.py
@@ -1,7 +1,17 @@
+import jwt
 from django.contrib.auth.backends import ModelBackend
 
 from ..account.models import User
-from ..graphql.account.dataloaders import load_user_from_request
+from ..graphql.account.dataloaders import UserByEmailLoader
+from .auth import get_token_from_request
+from .jwt import (
+    JWT_ACCESS_TYPE,
+    JWT_THIRDPARTY_ACCESS_TYPE,
+    PERMISSIONS_FIELD,
+    is_saleor_token,
+    jwt_decode,
+)
+from .permissions import get_permissions_from_codenames, get_permissions_from_names
 
 
 class JSONWebTokenBackend(ModelBackend):
@@ -46,3 +56,40 @@ class PluginBackend(JSONWebTokenBackend):
         if request is None:
             return None
         return request.plugins.authenticate_user(request)
+
+
+def load_user_from_request(request):
+    if request is None:
+        return None
+    jwt_token = get_token_from_request(request)
+    if not jwt_token or not is_saleor_token(jwt_token):
+        return None
+    payload = jwt_decode(jwt_token)
+
+    jwt_type = payload.get("type")
+    if jwt_type not in [JWT_ACCESS_TYPE, JWT_THIRDPARTY_ACCESS_TYPE]:
+        raise jwt.InvalidTokenError(
+            "Invalid token. Create new one by using tokenCreate mutation."
+        )
+    permissions = payload.get(PERMISSIONS_FIELD, None)
+
+    user = UserByEmailLoader(request).load(payload["email"]).get()
+    user_jwt_token = payload.get("token")
+    if not user_jwt_token or not user:
+        raise jwt.InvalidTokenError(
+            "Invalid token. Create new one by using tokenCreate mutation."
+        )
+    if user.jwt_token_key != user_jwt_token:
+        raise jwt.InvalidTokenError(
+            "Invalid token. Create new one by using tokenCreate mutation."
+        )
+
+    if permissions is not None:
+        token_permissions = get_permissions_from_names(permissions)
+        token_codenames = [perm.codename for perm in token_permissions]
+        user.effective_permissions = get_permissions_from_codenames(token_codenames)
+        user.is_staff = True if user.effective_permissions else False
+
+    if payload.get("is_staff"):
+        user.is_staff = True
+    return user
